+++ main/src/main/java/ru/practicum/ewm/model/events/dto/EventShortDto.java
@@ -18,7 +18,7 @@ public class EventShortDto {
     private String annotation; //Краткое описание
     @NotNull
     private CategoryDto category; //Категория
-    private Long confirmedRequests; //Количество одобренных заявок на участие в данном событии
+    private Integer confirmedRequests; //Количество одобренных заявок на участие в данном событии
     private String eventDate; //Дата и время на которые намечено событие (в формате "yyyy-MM-dd HH:mm:ss")
     private UserShortDto initiator; //Дата и время на которые намечено событие (в формате "yyyy-MM-dd HH:mm:ss")
     @NotNull
